+++ src/Maps/Projection.cpp
@@ -48,8 +48,8 @@ public:
 
     Coord mercatorInverse(const QPointF& point) const
     {
-        int longitude = angToCoord(point.x()*180.0/EQUATORIALMETERHALFCIRCUMFERENCE);
-        int latitude = radToCoord(atan(sinh(point.y()/EQUATORIALMETERHALFCIRCUMFERENCE*M_PI)));
+        double longitude = angToCoord(point.x()*180.0/EQUATORIALMETERHALFCIRCUMFERENCE);
+        double latitude = radToCoord(atan(sinh(point.y()/EQUATORIALMETERHALFCIRCUMFERENCE*M_PI)));
 
         return Coord(latitude, longitude);
     }
